requirements.txt
asgiref==3.11.0
crispy-bootstrap5==2025.6
dj-database-url==3.1.0
Django==6.0
django-cors-headers==4.9.0
django-crispy-forms==2.5
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.0.0
mysqlclient==2.2.7
packaging==26.0
psycopg2==2.9.11
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.5
tzdata==2025.3
